0.7.12.pcl-build1.2 	Bring src/pcl/low.lisp into the build 	... don't import SB-KERNEL:FUNCALLABLE-INSTANCE-P any more 	... move the !DEFSTRUCT-WITH-ALTERNATE-METACLASS calls into a 		file (target-low) that is built only by the 		cross-compiler, not the host compiler 	... usual package prefix frobs
This version builds and passes all its tests.  We aren't yet at the
"interesting" stage of this process, though; that comes up shortly, when
we try to migrate LOAD-DEFCLASS.

LOAD-DEFCLASS is much like %DEFSTRUCT, except that it gets redefined
halfway through PCL's build process as part of the bootstrap (in
src/pcl/std-class.lisp; see REAL-LOAD-DEFCLASS).

SBCL's way of dealing with these is probably instead to do fixups
(corresponding to PCL's braid?) at cold-init time, via !COLD-INIT-FORMS,
executed before the remaining warm load.  This may be the way to bridge
the gap between the early class information and the full PCL, or it may
not... we shall see.