0.7.12.pcl-build1.4: 	fngen only depends on low in the original defsys description 		of PCL, so 	... do trivial adjustments to move fngen into cold build
Sadly, that's the last easy case.

The obvious next step is to work on defclass.lisp until it yields, as it
is the next in the build order.  However, while eventually this is
likely to work, it's not going to advance the cause of rationalizing the
build; sure, we can build PCL under the host Lisp -- it was, after all,
called Portable Common LOOPS -- but that's not actually advancing the
cause of integrating it properly.

Maybe better would be to bring the braid into the cold build?  As part
of PCL's build, it makes PCL classes for every built-in-class.  We
_already_ have a big loop over built-in-classes at the start of
cold-init: maybe it would make sense to do the PCL classes there?