First attempt (non-working but ought to in principle, I think, probably) at implementing interrupt-threads using pthreads.
Two known issues:

1) The immediate and pressing reason it doesn't work is that something
in interrupt_thread is lopping off the high word of the return value
from alloc_cons before storing it in th->interrupt_fun.  This is
undoubtedy some problem with abuse of C's static typechecking, but I
cannot presently see what.  You can observe this with something as
simple as

(sb-thread:make-thread (lambda () (sleep 60)))
(sb-thread:interrupt-thread * #'break)

2) GC does not know about this thread->interrupt_fun slot and should
be told, in case it relocates the list between signal issuance (is
that a word?) and delivery.