1.0.27.33: make various form orders deterministic
Usually involves sorting the output of a hash-table loop or set
operation.

3 commit messages follow:

make the order of (setf cl:foo) defining forms deterministic

alphabetize the automated out-of-line definitions of modular functions

Otherwise we go in hash-table order, which is not noticeably the same
between clisp and sbcl hosts.

UNION can return entries in arbitrary order

So SORT [a copy of: don't mutate the source code!] the UNION of
signed-num and unsigned-num by symbol name.