1.0.27.34: various trivial fixups for portability / repeatability
The only one that is potentially controversial is the use of
READ-PRESERVING-WHITESPACE...

3 commit messages follow:

don't print array SB!KERNEL:TYPE in internal error strings

Use the specifier instead.  (This is a long-standing bug; FIXME: try to
find a test case).

Use read-preserving-whitespace rather than just read in the compiler

With just CL:READ, at least CLISP and SBCL differ on the source
locations dumped in the fasls; with READ-PRESERVING-WHITESPACE, things
are consistent.

disassembler / printer names.

The compiler wants to generate names based on all sorts of information,
including byte specs, and attempts to make those names by printing all
that information into one big string.  Unfortunately, that allows the
host to insert line breaks, which it will do with maximal perversity.
Bind *PRINT-RIGHT-MARGIN* around the printing call in an attempt to
minimize this problem.