1.0.27.36: host CL symbols with other home packages
Genesis already knew about the case of a symbol exported from the CL
package with a different home package.  For repeatable FASLs, the dumper
and the xref internals detector also need to know.

2 commit messages follow:

special case dumping of CL symbols with other home packages

Just like in genesis, we need to deal with CL symbols which are
permitted to have a home package that's not CL.  SBCL doesn't do that,
but other implementations legitimately can and do; nevertheless, dump
as though it were a CL symbol.

xref cross-compilation consistency fixes

Treat as internal symbols (a) symbols with home package being "SB-XC",
and (b) symbols which are external in the CL package but whose home
package is elsewhere.