1.0.27.38: make FAILED-AVER use a list form of code
Previously, we constructed a printed version of the code and used that,
but it seems remarkably hard to get identical printed contents from identical
list structure in three different implementations: indentation,
line breaks, QUOTE and FUNCTION, and so on all seem to vary.

2 previous commit messages follow:

bind printer control variables in FAILED-AVER

FAILED-AVER prints source code with ~A.  If printer control variables
are different in different implementations, then the error message will
be different.

Actually at the moment the binding (of *PRINT-PRETTY* to T) is probably
a no-op.  We tried binding *PRINT-PRETTY* to NIL to get the same output
as XCL, but apparently, other implementations (CLISP, reportedly ECL)
don't obey CLHS 22.1.3.5 for printing conses when the pretty printer is
off.

another attempt to tame AVER

binding printer control variables is all well and good, but linebreaks
cause problems.  We could probably deal with that with a suitable value
for *pprint-right-margin*, but... instead, just save the form, not its
printed representation.