1.0.32.31: type system now understands (and <array-type> (not simple-array))
Taken to mean the <array-type> with COMPLEXP T (rather than :MAYBE).
Adjust the type test transform to use the old technique for testing for
complex arrays (using an explicit (NOT SIMPLE-ARRAY) test rather than a
full call to %TYPEP, as you would otherwise get; this is a KLUDGE, but
no worse than before).

Include a test case for bug #309129, which this fixes.