1.0.33.15: preparation for UTF external formats
On input, there are potentially two non-constant sizes of data that need
to be read; for variable-width formats, there are both the number of bytes
for a particular character, and the number of bytes that need to be read in
order to know how many bytes need to be read for a character (previously
hardwired to 1).  Separate out these two sizes in preparation for UTF-16 and
UTF-32 external formats.