1.0.42.1: add release script
The script automates many of the tasks involved in performing the minimum
level of sanity checking that I am comfortable with, and enumerates others
that remain unscriptable despite 10 years of user interface "improvements"
at Sourceforge.  It is unlikely to generalise to other release managers'
setups and taste without a little effort.