cleanup DESCRIBE of symbols naming type specifiers a bit
Firstly, don't indent the "... names a type specifier".  But then
wander around in confusion for a while trying to debug where the extra
indentation comes from, before working out that the pprint-newline at
the end of describe-class introduces indentation as well as a newline,
causing the subsequent logical block to be opened at an unexpected
horizontal position.  Fix fix fix.